@@ -1,17 +1,19 @@
 [package]
 name = "alacritty_config"
-version = "0.1.2-dev"
+version = "0.2.1-dev"
 authors = ["Christian Duerr <contact@christianduerr.com>"]
-license = "MIT/Apache-2.0"
+license = "MIT OR Apache-2.0"
 description = "Alacritty configuration abstractions"
-homepage = "https://github.com/alacritty/alacritty"
+homepage = "https://alacritty.org"
+repository = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.60.0"
+rust-version = "1.70.0"
 
 [dependencies]
 log = { version = "0.4.17", features = ["serde"] }
-serde_yaml = "0.8.24"
-serde = "1.0.137"
+serde = "1.0.163"
+toml = "0.8.2"
 
-[target.'cfg(target_os = "macos")'.dependencies]
-winit = { version = "0.28.2", default-features = false, features = ["serde"] }
+[dev-dependencies]
+alacritty_config_derive = { version = "0.2.3-dev", path = "../alacritty_config_derive" }
+serde = { version = "1.0.163", features = ["derive"] }
