@@ -1,49 +1,44 @@
 [package]
 name = "alacritty_terminal"
-version = "0.18.1-dev"
+version = "0.23.0-dev"
 authors = ["Christian Duerr <contact@christianduerr.com>", "Joe Wilm <joe@jwilm.com>"]
 license = "Apache-2.0"
 description = "Library for writing terminal emulators"
 readme = "../README.md"
-homepage = "https://github.com/alacritty/alacritty"
+homepage = "https://alacritty.org"
+repository = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.60.0"
+rust-version = "1.70.0"
 
-[dependencies.alacritty_config_derive]
-path = "../alacritty_config_derive"
-version = "0.2.2-dev"
-
-[dependencies.alacritty_config]
-path = "../alacritty_config"
-version = "0.1.2-dev"
+[features]
+default = ["serde"]
+serde = ["dep:serde", "bitflags/serde", "vte/serde"]
 
 [dependencies]
+base64 = "0.22.0"
+bitflags = "2.4.1"
+home = "0.5.5"
 libc = "0.2"
-bitflags = "1"
-parking_lot = "0.12.0"
-serde = { version = "1", features = ["derive", "rc"] }
-serde_yaml = "0.8"
-vte = { version = "0.10.0", default-features = false }
-mio = "0.6.20"
-mio-extras = "2"
 log = "0.4"
+parking_lot = "0.12.0"
+polling = "3.0.0"
+regex-automata = "0.4.3"
 unicode-width = "0.1"
-base64 = "0.13.0"
-regex-automata = "0.1.9"
-dirs = "4.0.0"
+vte = { version = "0.13.0", default-features = false, features = ["ansi", "serde"] }
+serde = { version = "1", features = ["derive", "rc"], optional = true }
 
 [target.'cfg(unix)'.dependencies]
-nix = "0.24.1"
+rustix-openpty = "0.1.1"
 signal-hook = "0.3.10"
-signal-hook-mio = { version = "0.2.1", features = ["support-v0_6"] }
 
 [target.'cfg(windows)'.dependencies]
-miow = "0.3"
-mio-anonymous-pipes = "0.2"
-windows-sys = { version = "0.36", features = [
+piper = "0.2.1"
+miow = "0.6.0"
+windows-sys = { version = "0.48.0", features = [
     "Win32_System_Console",
     "Win32_Foundation",
     "Win32_Security",
+    "Win32_System_LibraryLoader",
     "Win32_System_Threading",
     "Win32_System_WindowsProgramming",
 ]}
