@@ -15,12 +15,15 @@ fn main() {
     let dest = env::var("OUT_DIR").unwrap();
     let mut file = File::create(Path::new(&dest).join("gl_bindings.rs")).unwrap();
 
-    Registry::new(Api::Gl, (3, 3), Profile::Core, Fallbacks::All, ["GL_ARB_blend_func_extended"])
-        .write_bindings(GlobalGenerator, &mut file)
-        .unwrap();
+    Registry::new(Api::Gl, (3, 3), Profile::Core, Fallbacks::All, [
+        "GL_ARB_blend_func_extended",
+        "GL_KHR_debug",
+    ])
+    .write_bindings(GlobalGenerator, &mut file)
+    .unwrap();
 
     #[cfg(windows)]
-    embed_resource::compile("./windows/alacritty.rc");
+    embed_resource::compile("./windows/alacritty.rc", embed_resource::NONE);
 }
 
 fn commit_hash() -> Option<String> {
