@@ -2,7 +2,8 @@ use serde::de::Error as SerdeError;
 use serde::{Deserialize, Deserializer};
 
 use alacritty_config_derive::ConfigDeserialize;
-use alacritty_terminal::term::color::{CellRgb, Rgb};
+
+use crate::display::color::{CellRgb, Rgb};
 
 #[derive(ConfigDeserialize, Clone, Debug, Default, PartialEq, Eq)]
 pub struct Colors {
@@ -18,16 +19,17 @@ pub struct Colors {
     pub line_indicator: LineIndicatorColors,
     pub hints: HintColors,
     pub transparent_background_colors: bool,
+    pub draw_bold_text_with_bright_colors: bool,
     footer_bar: BarColors,
 }
 
 impl Colors {
     pub fn footer_bar_foreground(&self) -> Rgb {
-        self.search.bar.foreground.or(self.footer_bar.foreground).unwrap_or(self.primary.background)
+        self.footer_bar.foreground.unwrap_or(self.primary.background)
     }
 
     pub fn footer_bar_background(&self) -> Rgb {
-        self.search.bar.background.or(self.footer_bar.background).unwrap_or(self.primary.foreground)
+        self.footer_bar.background.unwrap_or(self.primary.foreground)
     }
 }
 
@@ -52,8 +54,8 @@ pub struct HintStartColors {
 impl Default for HintStartColors {
     fn default() -> Self {
         Self {
-            foreground: CellRgb::Rgb(Rgb { r: 0x1d, g: 0x1f, b: 0x21 }),
-            background: CellRgb::Rgb(Rgb { r: 0xe9, g: 0xff, b: 0x5e }),
+            foreground: CellRgb::Rgb(Rgb::new(0x18, 0x18, 0x18)),
+            background: CellRgb::Rgb(Rgb::new(0xf4, 0xbf, 0x75)),
         }
     }
 }
@@ -67,13 +69,14 @@ pub struct HintEndColors {
 impl Default for HintEndColors {
     fn default() -> Self {
         Self {
-            foreground: CellRgb::Rgb(Rgb { r: 0xe9, g: 0xff, b: 0x5e }),
-            background: CellRgb::Rgb(Rgb { r: 0x1d, g: 0x1f, b: 0x21 }),
+            foreground: CellRgb::Rgb(Rgb::new(0x18, 0x18, 0x18)),
+            background: CellRgb::Rgb(Rgb::new(0xac, 0x42, 0x42)),
         }
     }
 }
 
 #[derive(Deserialize, Copy, Clone, Default, Debug, PartialEq, Eq)]
+#[serde(deny_unknown_fields)]
 pub struct IndexedColor {
     pub color: Rgb,
 
@@ -126,8 +129,6 @@ impl Default for InvertedCellColors {
 pub struct SearchColors {
     pub focused_match: FocusedMatchColors,
     pub matches: MatchColors,
-    #[config(deprecated = "use `colors.footer_bar` instead")]
-    bar: BarColors,
 }
 
 #[derive(ConfigDeserialize, Debug, Copy, Clone, PartialEq, Eq)]
@@ -139,8 +140,8 @@ pub struct FocusedMatchColors {
 impl Default for FocusedMatchColors {
     fn default() -> Self {
         Self {
-            background: CellRgb::Rgb(Rgb { r: 0x00, g: 0x00, b: 0x00 }),
-            foreground: CellRgb::Rgb(Rgb { r: 0xff, g: 0xff, b: 0xff }),
+            background: CellRgb::Rgb(Rgb::new(0xf4, 0xbf, 0x75)),
+            foreground: CellRgb::Rgb(Rgb::new(0x18, 0x18, 0x18)),
         }
     }
 }
@@ -154,8 +155,8 @@ pub struct MatchColors {
 impl Default for MatchColors {
     fn default() -> Self {
         Self {
-            background: CellRgb::Rgb(Rgb { r: 0xff, g: 0xff, b: 0xff }),
-            foreground: CellRgb::Rgb(Rgb { r: 0x00, g: 0x00, b: 0x00 }),
+            background: CellRgb::Rgb(Rgb::new(0xac, 0x42, 0x42)),
+            foreground: CellRgb::Rgb(Rgb::new(0x18, 0x18, 0x18)),
         }
     }
 }
@@ -177,8 +178,8 @@ pub struct PrimaryColors {
 impl Default for PrimaryColors {
     fn default() -> Self {
         PrimaryColors {
-            background: Rgb { r: 0x1d, g: 0x1f, b: 0x21 },
-            foreground: Rgb { r: 0xc5, g: 0xc8, b: 0xc6 },
+            background: Rgb::new(0x18, 0x18, 0x18),
+            foreground: Rgb::new(0xd8, 0xd8, 0xd8),
             bright_foreground: Default::default(),
             dim_foreground: Default::default(),
         }
@@ -200,14 +201,14 @@ pub struct NormalColors {
 impl Default for NormalColors {
     fn default() -> Self {
         NormalColors {
-            black: Rgb { r: 0x1d, g: 0x1f, b: 0x21 },
-            red: Rgb { r: 0xcc, g: 0x66, b: 0x66 },
-            green: Rgb { r: 0xb5, g: 0xbd, b: 0x68 },
-            yellow: Rgb { r: 0xf0, g: 0xc6, b: 0x74 },
-            blue: Rgb { r: 0x81, g: 0xa2, b: 0xbe },
-            magenta: Rgb { r: 0xb2, g: 0x94, b: 0xbb },
-            cyan: Rgb { r: 0x8a, g: 0xbe, b: 0xb7 },
-            white: Rgb { r: 0xc5, g: 0xc8, b: 0xc6 },
+            black: Rgb::new(0x18, 0x18, 0x18),
+            red: Rgb::new(0xac, 0x42, 0x42),
+            green: Rgb::new(0x90, 0xa9, 0x59),
+            yellow: Rgb::new(0xf4, 0xbf, 0x75),
+            blue: Rgb::new(0x6a, 0x9f, 0xb5),
+            magenta: Rgb::new(0xaa, 0x75, 0x9f),
+            cyan: Rgb::new(0x75, 0xb5, 0xaa),
+            white: Rgb::new(0xd8, 0xd8, 0xd8),
         }
     }
 }
@@ -226,15 +227,18 @@ pub struct BrightColors {
 
 impl Default for BrightColors {
     fn default() -> Self {
+        // Generated with oklab by multiplying brightness by 1.12 and then adjusting numbers
+        // to make them look "nicer". Yellow color was generated the same way, however the first
+        // srgb representable color was picked.
         BrightColors {
-            black: Rgb { r: 0x66, g: 0x66, b: 0x66 },
-            red: Rgb { r: 0xd5, g: 0x4e, b: 0x53 },
-            green: Rgb { r: 0xb9, g: 0xca, b: 0x4a },
-            yellow: Rgb { r: 0xe7, g: 0xc5, b: 0x47 },
-            blue: Rgb { r: 0x7a, g: 0xa6, b: 0xda },
-            magenta: Rgb { r: 0xc3, g: 0x97, b: 0xd8 },
-            cyan: Rgb { r: 0x70, g: 0xc0, b: 0xb1 },
-            white: Rgb { r: 0xea, g: 0xea, b: 0xea },
+            black: Rgb::new(0x6b, 0x6b, 0x6b),
+            red: Rgb::new(0xc5, 0x55, 0x55),
+            green: Rgb::new(0xaa, 0xc4, 0x74),
+            yellow: Rgb::new(0xfe, 0xca, 0x88),
+            blue: Rgb::new(0x82, 0xb8, 0xc8),
+            magenta: Rgb::new(0xc2, 0x8c, 0xb8),
+            cyan: Rgb::new(0x93, 0xd3, 0xc3),
+            white: Rgb::new(0xf8, 0xf8, 0xf8),
         }
     }
 }
@@ -253,15 +257,16 @@ pub struct DimColors {
 
 impl Default for DimColors {
     fn default() -> Self {
+        // Generated with builtin alacritty's color dimming function.
         DimColors {
-            black: Rgb { r: 0x13, g: 0x14, b: 0x15 },
-            red: Rgb { r: 0x86, g: 0x43, b: 0x43 },
-            green: Rgb { r: 0x77, g: 0x7c, b: 0x44 },
-            yellow: Rgb { r: 0x9e, g: 0x82, b: 0x4c },
-            blue: Rgb { r: 0x55, g: 0x6a, b: 0x7d },
-            magenta: Rgb { r: 0x75, g: 0x61, b: 0x7b },
-            cyan: Rgb { r: 0x5b, g: 0x7d, b: 0x78 },
-            white: Rgb { r: 0x82, g: 0x84, b: 0x82 },
+            black: Rgb::new(0x0f, 0x0f, 0x0f),
+            red: Rgb::new(0x71, 0x2b, 0x2b),
+            green: Rgb::new(0x5f, 0x6f, 0x3a),
+            yellow: Rgb::new(0xa1, 0x7e, 0x4d),
+            blue: Rgb::new(0x45, 0x68, 0x77),
+            magenta: Rgb::new(0x70, 0x4d, 0x68),
+            cyan: Rgb::new(0x4d, 0x77, 0x70),
+            white: Rgb::new(0x8e, 0x8e, 0x8e),
         }
     }
 }
