@@ -1,7 +1,5 @@
 use log::LevelFilter;
 
-use serde::Deserialize;
-
 use alacritty_config_derive::ConfigDeserialize;
 
 /// Debugging options.
@@ -23,6 +21,9 @@ pub struct Debug {
     /// The renderer alacritty should be using.
     pub renderer: Option<RendererPreference>,
 
+    /// Use EGL as display API if the current platform allows it.
+    pub prefer_egl: bool,
+
     /// Record ref test.
     #[config(skip)]
     pub ref_test: bool,
@@ -38,22 +39,20 @@ impl Default for Debug {
             highlight_damage: Default::default(),
             ref_test: Default::default(),
             renderer: Default::default(),
+            prefer_egl: Default::default(),
         }
     }
 }
 
 /// The renderer configuration options.
-#[derive(Deserialize, Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(ConfigDeserialize, Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub enum RendererPreference {
     /// OpenGL 3.3 renderer.
-    #[serde(rename = "glsl3")]
     Glsl3,
 
     /// GLES 2 renderer, with optional extensions like dual source blending.
-    #[serde(rename = "gles2")]
     Gles2,
 
     /// Pure GLES 2 renderer.
-    #[serde(rename = "gles2_pure")]
     Gles2Pure,
 }
