@@ -1,19 +1,17 @@
 use std::fmt::{self, Formatter};
-use std::os::raw::c_ulong;
 
 use log::{error, warn};
 use serde::de::{self, MapAccess, Visitor};
 use serde::{Deserialize, Deserializer, Serialize};
-use winit::window::{Fullscreen, Theme};
 
 #[cfg(target_os = "macos")]
-use winit::platform::macos::OptionAsAlt;
+use winit::platform::macos::OptionAsAlt as WinitOptionAsAlt;
+use winit::window::{Fullscreen, Theme as WinitTheme};
 
 use alacritty_config_derive::{ConfigDeserialize, SerdeReplace};
-use alacritty_terminal::config::{Percentage, LOG_TARGET_CONFIG};
-use alacritty_terminal::index::Column;
 
-use crate::config::ui_config::Delta;
+use crate::config::ui_config::{Delta, Percentage};
+use crate::config::LOG_TARGET_CONFIG;
 
 /// Default Alacritty name, used for window title and class.
 pub const DEFAULT_NAME: &str = "Alacritty";
@@ -31,10 +29,7 @@ pub struct WindowConfig {
 
     /// XEmbed parent.
     #[config(skip)]
-    pub embed: Option<c_ulong>,
-
-    /// System decorations theme variant.
-    pub decorations_theme_variant: Option<Theme>,
+    pub embed: Option<u32>,
 
     /// Spread out additional padding evenly.
     pub dynamic_padding: bool,
@@ -49,36 +44,41 @@ pub struct WindowConfig {
     /// Background opacity from 0.0 to 1.0.
     pub opacity: Percentage,
 
+    /// Request blur behind the window.
+    pub blur: bool,
+
     /// Controls which `Option` key should be treated as `Alt`.
-    #[cfg(target_os = "macos")]
-    pub option_as_alt: OptionAsAlt,
+    option_as_alt: OptionAsAlt,
 
     /// Resize increments.
     pub resize_increments: bool,
 
     /// Pixel padding.
-    padding: Delta<u8>,
+    padding: Delta<u16>,
 
     /// Initial dimensions.
     dimensions: Dimensions,
+
+    /// System decorations theme variant.
+    decorations_theme_variant: Option<Theme>,
 }
 
 impl Default for WindowConfig {
     fn default() -> Self {
         Self {
             dynamic_title: true,
+            blur: Default::default(),
+            embed: Default::default(),
+            padding: Default::default(),
+            opacity: Default::default(),
             position: Default::default(),
+            identity: Default::default(),
+            dimensions: Default::default(),
             decorations: Default::default(),
             startup_mode: Default::default(),
-            embed: Default::default(),
-            decorations_theme_variant: Default::default(),
             dynamic_padding: Default::default(),
-            identity: Identity::default(),
-            opacity: Default::default(),
-            padding: Default::default(),
-            dimensions: Default::default(),
             resize_increments: Default::default(),
-            #[cfg(target_os = "macos")]
+            decorations_theme_variant: Default::default(),
             option_as_alt: Default::default(),
         }
     }
@@ -87,7 +87,7 @@ impl Default for WindowConfig {
 impl WindowConfig {
     #[inline]
     pub fn dimensions(&self) -> Option<Dimensions> {
-        let (lines, columns) = (self.dimensions.lines, self.dimensions.columns.0);
+        let (lines, columns) = (self.dimensions.lines, self.dimensions.columns);
         let (lines_is_non_zero, columns_is_non_zero) = (lines != 0, columns != 0);
 
         if lines_is_non_zero && columns_is_non_zero {
@@ -137,6 +137,20 @@ impl WindowConfig {
     pub fn maximized(&self) -> bool {
         self.startup_mode == StartupMode::Maximized
     }
+
+    #[cfg(target_os = "macos")]
+    pub fn option_as_alt(&self) -> WinitOptionAsAlt {
+        match self.option_as_alt {
+            OptionAsAlt::OnlyLeft => WinitOptionAsAlt::OnlyLeft,
+            OptionAsAlt::OnlyRight => WinitOptionAsAlt::OnlyRight,
+            OptionAsAlt::Both => WinitOptionAsAlt::Both,
+            OptionAsAlt::None => WinitOptionAsAlt::None,
+        }
+    }
+
+    pub fn theme(&self) -> Option<WinitTheme> {
+        self.decorations_theme_variant.map(WinitTheme::from)
+    }
 }
 
 #[derive(ConfigDeserialize, Debug, Clone, PartialEq, Eq)]
@@ -154,35 +168,22 @@ impl Default for Identity {
     }
 }
 
-#[derive(ConfigDeserialize, Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(ConfigDeserialize, Default, Debug, Copy, Clone, PartialEq, Eq)]
 pub enum StartupMode {
+    #[default]
     Windowed,
     Maximized,
     Fullscreen,
-    #[cfg(target_os = "macos")]
     SimpleFullscreen,
 }
 
-impl Default for StartupMode {
-    fn default() -> StartupMode {
-        StartupMode::Windowed
-    }
-}
-
-#[derive(ConfigDeserialize, Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(ConfigDeserialize, Default, Debug, Copy, Clone, PartialEq, Eq)]
 pub enum Decorations {
+    #[default]
     Full,
-    #[cfg(target_os = "macos")]
     Transparent,
-    #[cfg(target_os = "macos")]
     Buttonless,
     None,
-}
-
-impl Default for Decorations {
-    fn default() -> Decorations {
-        Decorations::Full
-    }
 }
 
 /// Window Dimensions.
@@ -191,7 +192,7 @@ impl Default for Decorations {
 #[derive(ConfigDeserialize, Default, Debug, Copy, Clone, PartialEq, Eq)]
 pub struct Dimensions {
     /// Window width in character columns.
-    pub columns: Column,
+    pub columns: usize,
 
     /// Window Height in character lines.
     pub lines: usize,
@@ -242,7 +243,7 @@ impl<'de> Deserialize<'de> for Class {
             {
                 let mut class = Self::Value::default();
 
-                while let Some((key, value)) = map.next_entry::<String, serde_yaml::Value>()? {
+                while let Some((key, value)) = map.next_entry::<String, toml::Value>()? {
                     match key.as_str() {
                         "instance" => match String::deserialize(value) {
                             Ok(instance) => class.instance = instance,
@@ -262,7 +263,7 @@ impl<'de> Deserialize<'de> for Class {
                                 );
                             },
                         },
-                        _ => (),
+                        key => warn!(target: LOG_TARGET_CONFIG, "Unrecognized class field: {key}"),
                     }
                 }
 
@@ -271,5 +272,37 @@ impl<'de> Deserialize<'de> for Class {
         }
 
         deserializer.deserialize_any(ClassVisitor)
+    }
+}
+
+#[derive(ConfigDeserialize, Default, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum OptionAsAlt {
+    /// The left `Option` key is treated as `Alt`.
+    OnlyLeft,
+
+    /// The right `Option` key is treated as `Alt`.
+    OnlyRight,
+
+    /// Both `Option` keys are treated as `Alt`.
+    Both,
+
+    /// No special handling is applied for `Option` key.
+    #[default]
+    None,
+}
+
+/// System decorations theme variant.
+#[derive(ConfigDeserialize, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Theme {
+    Light,
+    Dark,
+}
+
+impl From<Theme> for WinitTheme {
+    fn from(theme: Theme) -> Self {
+        match theme {
+            Theme::Light => WinitTheme::Light,
+            Theme::Dark => WinitTheme::Dark,
+        }
     }
 }
