@@ -1,9 +1,6 @@
-#[rustfmt::skip]
-#[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-use {
-    wayland_client::protocol::wl_surface::WlSurface,
-    wayland_client::{Attached, EventQueue, Proxy},
-    winit::platform::wayland::{EventLoopWindowTargetExtWayland, WindowExtWayland},
+#[cfg(not(any(target_os = "macos", windows)))]
+use winit::platform::startup_notify::{
+    self, EventLoopExtStartupNotify, WindowBuilderExtStartupNotify,
 };
 
 #[cfg(all(not(feature = "x11"), not(any(target_os = "macos", windows))))]
@@ -13,17 +10,13 @@ use winit::platform::wayland::WindowBuilderExtWayland;
 #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
 use {
     std::io::Cursor,
-
-    winit::platform::x11::{WindowExtX11, WindowBuilderExtX11},
+    winit::platform::x11::{WindowBuilderExtX11, EventLoopWindowTargetExtX11},
     glutin::platform::x11::X11VisualInfo,
-    x11_dl::xlib::{Display as XDisplay, PropModeReplace, XErrorEvent, Xlib},
     winit::window::Icon,
     png::Decoder,
 };
 
 use std::fmt::{self, Display, Formatter};
-use std::sync::atomic::AtomicBool;
-use std::sync::Arc;
 
 #[cfg(target_os = "macos")]
 use {
@@ -34,15 +27,14 @@ use {
 };
 
 use raw_window_handle::{HasRawWindowHandle, RawWindowHandle};
-
 use winit::dpi::{PhysicalPosition, PhysicalSize};
 use winit::event_loop::EventLoopWindowTarget;
 use winit::monitor::MonitorHandle;
 #[cfg(windows)]
 use winit::platform::windows::IconExtWindows;
 use winit::window::{
-    CursorIcon, Fullscreen, ImePurpose, UserAttentionType, Window as WinitWindow, WindowBuilder,
-    WindowId,
+    CursorIcon, Fullscreen, ImePurpose, Theme, UserAttentionType, Window as WinitWindow,
+    WindowBuilder, WindowId,
 };
 
 use alacritty_terminal::index::Point;
@@ -107,20 +99,20 @@ impl From<crossfont::Error> for Error {
 /// Wraps the underlying windowing library to provide a stable API in Alacritty.
 pub struct Window {
     /// Flag tracking that we have a frame we can draw.
-    pub has_frame: Arc<AtomicBool>,
-
-    /// Attached Wayland surface to request new frame events.
-    #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-    pub wayland_surface: Option<Attached<WlSurface>>,
+    pub has_frame: bool,
 
     /// Cached scale factor for quickly scaling pixel sizes.
     pub scale_factor: f64,
+
+    /// Flag indicating whether redraw was requested.
+    pub requested_redraw: bool,
 
     window: WinitWindow,
 
     /// Current window title.
     title: String,
 
+    is_x11: bool,
     current_mouse_cursor: CursorIcon,
     mouse_visible: bool,
 }
@@ -133,8 +125,9 @@ impl Window {
         event_loop: &EventLoopWindowTarget<E>,
         config: &UiConfig,
         identity: &Identity,
-        #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-        wayland_event_queue: Option<&EventQueue>,
+        #[rustfmt::skip]
+        #[cfg(target_os = "macos")]
+        tabbing_id: &Option<String>,
         #[rustfmt::skip]
         #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
         x11_visual: Option<X11VisualInfo>,
@@ -145,6 +138,8 @@ impl Window {
             &config.window,
             #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
             x11_visual,
+            #[cfg(target_os = "macos")]
+            tabbing_id,
         );
 
         if let Some(position) = config.window.position {
@@ -152,20 +147,31 @@ impl Window {
                 .with_position(PhysicalPosition::<i32>::from((position.x, position.y)));
         }
 
+        #[cfg(not(any(target_os = "macos", windows)))]
+        if let Some(token) = event_loop.read_token_from_env() {
+            log::debug!("Activating window with token: {token:?}");
+            window_builder = window_builder.with_activation_token(token);
+
+            // Remove the token from the env.
+            startup_notify::reset_activation_token_env();
+        }
+
+        // On X11, embed the window inside another if the parent ID has been set.
+        #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
+        if let Some(parent_window_id) = event_loop.is_x11().then_some(config.window.embed).flatten()
+        {
+            window_builder = window_builder.with_embed_parent_window(parent_window_id);
+        }
+
         let window = window_builder
             .with_title(&identity.title)
-            .with_theme(config.window.decorations_theme_variant)
+            .with_theme(config.window.theme())
             .with_visible(false)
             .with_transparent(true)
+            .with_blur(config.window.blur)
             .with_maximized(config.window.maximized())
             .with_fullscreen(config.window.fullscreen())
             .build(event_loop)?;
-
-        // Check if we're running Wayland to disable vsync.
-        #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-        let is_wayland = event_loop.is_wayland();
-        #[cfg(all(not(feature = "wayland"), not(any(target_os = "macos", windows))))]
-        let is_wayland = false;
 
         // Text cursor.
         let current_mouse_cursor = CursorIcon::Text;
@@ -181,36 +187,19 @@ impl Window {
         #[cfg(target_os = "macos")]
         use_srgb_color_space(&window);
 
-        #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
-        if !is_wayland {
-            // On X11, embed the window inside another if the parent ID has been set.
-            if let Some(parent_window_id) = config.window.embed {
-                x_embed_window(&window, parent_window_id);
-            }
-        }
-
-        #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-        let wayland_surface = if is_wayland {
-            // Attach surface to Alacritty's internal wayland queue to handle frame callbacks.
-            let surface = window.wayland_surface().unwrap();
-            let proxy: Proxy<WlSurface> = unsafe { Proxy::from_c_ptr(surface as _) };
-            Some(proxy.attach(wayland_event_queue.as_ref().unwrap().token()))
-        } else {
-            None
-        };
-
         let scale_factor = window.scale_factor();
         log::info!("Window scale factor: {}", scale_factor);
+        let is_x11 = matches!(window.raw_window_handle(), RawWindowHandle::Xlib(_));
 
         Ok(Self {
+            requested_redraw: false,
+            title: identity.title,
             current_mouse_cursor,
             mouse_visible: true,
-            window,
-            title: identity.title,
-            has_frame: Arc::new(AtomicBool::new(true)),
-            #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-            wayland_surface,
+            has_frame: true,
             scale_factor,
+            window,
+            is_x11,
         })
     }
 
@@ -220,8 +209,8 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_inner_size(&self, size: PhysicalSize<u32>) {
-        self.window.set_inner_size(size);
+    pub fn request_inner_size(&self, size: PhysicalSize<u32>) {
+        let _ = self.window.request_inner_size(size);
     }
 
     #[inline]
@@ -248,8 +237,11 @@ impl Window {
     }
 
     #[inline]
-    pub fn request_redraw(&self) {
-        self.window.request_redraw();
+    pub fn request_redraw(&mut self) {
+        if !self.requested_redraw {
+            self.requested_redraw = true;
+            self.window.request_redraw();
+        }
     }
 
     #[inline]
@@ -296,7 +288,7 @@ impl Window {
 
         #[cfg(feature = "x11")]
         let builder = match x11_visual {
-            Some(visual) => builder.with_x11_visual(visual.into_raw()),
+            Some(visual) => builder.with_x11_visual(visual.visual_id() as u32),
             None => builder,
         };
 
@@ -313,8 +305,16 @@ impl Window {
     }
 
     #[cfg(target_os = "macos")]
-    pub fn get_platform_window(_: &Identity, window_config: &WindowConfig) -> WindowBuilder {
-        let window = WindowBuilder::new().with_option_as_alt(window_config.option_as_alt);
+    pub fn get_platform_window(
+        _: &Identity,
+        window_config: &WindowConfig,
+        tabbing_id: &Option<String>,
+    ) -> WindowBuilder {
+        let mut window = WindowBuilder::new().with_option_as_alt(window_config.option_as_alt());
+
+        if let Some(tabbing_id) = tabbing_id {
+            window = window.with_tabbing_identifier(tabbing_id);
+        }
 
         match window_config.decorations {
             Decorations::Full => window,
@@ -345,6 +345,10 @@ impl Window {
         self.window.set_transparent(transparent);
     }
 
+    pub fn set_blur(&self, blur: bool) {
+        self.window.set_blur(blur);
+    }
+
     pub fn set_maximized(&self, maximized: bool) {
         self.window.set_maximized(maximized);
     }
@@ -365,6 +369,17 @@ impl Window {
     /// Toggle the window's maximized state.
     pub fn toggle_maximized(&self) {
         self.set_maximized(!self.window.is_maximized());
+    }
+
+    /// Inform windowing system about presenting to the window.
+    ///
+    /// Should be called right before presenting to the window with e.g. `eglSwapBuffers`.
+    pub fn pre_present_notify(&self) {
+        self.window.pre_present_notify();
+    }
+
+    pub fn set_theme(&self, theme: Option<Theme>) {
+        self.window.set_theme(theme);
     }
 
     #[cfg(target_os = "macos")]
@@ -394,21 +409,32 @@ impl Window {
         self.window.set_simple_fullscreen(simple_fullscreen);
     }
 
-    #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-    pub fn wayland_surface(&self) -> Option<&Attached<WlSurface>> {
-        self.wayland_surface.as_ref()
-    }
-
     pub fn set_ime_allowed(&self, allowed: bool) {
-        self.window.set_ime_allowed(allowed);
+        // Skip runtime IME manipulation on X11 since it breaks some IMEs.
+        if !self.is_x11 {
+            self.window.set_ime_allowed(allowed);
+        }
     }
 
     /// Adjust the IME editor position according to the new location of the cursor.
     pub fn update_ime_position(&self, point: Point<usize>, size: &SizeInfo) {
+        // NOTE: X11 doesn't support cursor area, so we need to offset manually to not obscure
+        // the text.
+        let offset = if self.is_x11 { 1 } else { 0 };
         let nspot_x = f64::from(size.padding_x() + point.column.0 as f32 * size.cell_width());
-        let nspot_y = f64::from(size.padding_y() + (point.line + 1) as f32 * size.cell_height());
+        let nspot_y =
+            f64::from(size.padding_y() + (point.line + offset) as f32 * size.cell_height());
 
-        self.window.set_ime_position(PhysicalPosition::new(nspot_x, nspot_y));
+        // NOTE: some compositors don't like excluding too much and try to render popup at the
+        // bottom right corner of the provided area, so exclude just the full-width char to not
+        // obscure the cursor and not render popup at the end of the window.
+        let width = size.cell_width() as f64 * 2.;
+        let height = size.cell_height as f64;
+
+        self.window.set_ime_cursor_area(
+            PhysicalPosition::new(nspot_x, nspot_y),
+            PhysicalSize::new(width, height),
+        );
     }
 
     /// Disable macOS window shadows.
@@ -426,39 +452,34 @@ impl Window {
             let _: id = msg_send![raw_window, setHasShadow: value];
         }
     }
-}
 
-#[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
-fn x_embed_window(window: &WinitWindow, parent_id: std::os::raw::c_ulong) {
-    let (xlib_display, xlib_window) = match (window.xlib_display(), window.xlib_window()) {
-        (Some(display), Some(window)) => (display, window),
-        _ => return,
-    };
+    /// Select tab at the given `index`.
+    #[cfg(target_os = "macos")]
+    pub fn select_tab_at_index(&self, index: usize) {
+        self.window.select_tab_at_index(index);
+    }
 
-    let xlib = Xlib::open().expect("get xlib");
+    /// Select the last tab.
+    #[cfg(target_os = "macos")]
+    pub fn select_last_tab(&self) {
+        self.window.select_tab_at_index(self.window.num_tabs() - 1);
+    }
 
-    unsafe {
-        let atom = (xlib.XInternAtom)(xlib_display as *mut _, "_XEMBED".as_ptr() as *const _, 0);
-        (xlib.XChangeProperty)(
-            xlib_display as _,
-            xlib_window as _,
-            atom,
-            atom,
-            32,
-            PropModeReplace,
-            [0, 1].as_ptr(),
-            2,
-        );
+    /// Select next tab.
+    #[cfg(target_os = "macos")]
+    pub fn select_next_tab(&self) {
+        self.window.select_next_tab();
+    }
 
-        // Register new error handler.
-        let old_handler = (xlib.XSetErrorHandler)(Some(xembed_error_handler));
+    /// Select previous tab.
+    #[cfg(target_os = "macos")]
+    pub fn select_previous_tab(&self) {
+        self.window.select_previous_tab();
+    }
 
-        // Check for the existence of the target before attempting reparenting.
-        (xlib.XReparentWindow)(xlib_display as _, xlib_window as _, parent_id, 0, 0);
-
-        // Drain errors and restore original error handler.
-        (xlib.XSync)(xlib_display as _, 0);
-        (xlib.XSetErrorHandler)(old_handler);
+    #[cfg(target_os = "macos")]
+    pub fn tabbing_id(&self) -> String {
+        self.window.tabbing_identifier()
     }
 }
 
@@ -472,10 +493,4 @@ fn use_srgb_color_space(window: &WinitWindow) {
     unsafe {
         let _: () = msg_send![raw_window, setColorSpace: NSColorSpace::sRGBColorSpace(nil)];
     }
-}
-
-#[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
-unsafe extern "C" fn xembed_error_handler(_: *mut XDisplay, _: *mut XErrorEvent) -> i32 {
-    log::error!("Could not embed into specified window.");
-    std::process::exit(1);
 }
