@@ -7,25 +7,19 @@ use std::mem;
 #[cfg(not(windows))]
 use std::os::unix::io::{AsRawFd, RawFd};
 use std::rc::Rc;
-use std::sync::atomic::Ordering;
 use std::sync::Arc;
 
-use crossfont::Size;
 use glutin::config::GetGlConfig;
-use glutin::context::NotCurrentContext;
 use glutin::display::GetGlDisplay;
 #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
 use glutin::platform::x11::X11GlConfigExt;
-use log::{error, info};
+use log::info;
 use raw_window_handle::HasRawDisplayHandle;
 use serde_json as json;
-#[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-use wayland_client::EventQueue;
-use winit::event::{Event as WinitEvent, ModifiersState, WindowEvent};
+use winit::event::{Event as WinitEvent, Modifiers, WindowEvent};
 use winit::event_loop::{EventLoopProxy, EventLoopWindowTarget};
 use winit::window::WindowId;
 
-use alacritty_config::SerdeReplace;
 use alacritty_terminal::event::Event as TerminalEvent;
 use alacritty_terminal::event_loop::{EventLoop as PtyEventLoop, Msg, Notifier};
 use alacritty_terminal::grid::{Dimensions, Scroll};
@@ -35,14 +29,15 @@ use alacritty_terminal::term::test::TermSize;
 use alacritty_terminal::term::{Term, TermMode};
 use alacritty_terminal::tty;
 
-#[cfg(unix)]
-use crate::cli::IpcConfig;
-use crate::cli::WindowOptions;
+use crate::cli::{ParsedOptions, WindowOptions};
 use crate::clipboard::Clipboard;
 use crate::config::UiConfig;
 use crate::display::window::Window;
 use crate::display::Display;
-use crate::event::{ActionContext, Event, EventProxy, EventType, Mouse, SearchState, TouchPurpose};
+use crate::event::{
+    ActionContext, Event, EventProxy, InlineSearchState, Mouse, SearchState, TouchPurpose,
+};
+#[cfg(unix)]
 use crate::logging::LOG_TARGET_IPC_CONFIG;
 use crate::message_bar::MessageBuffer;
 use crate::scheduler::Scheduler;
@@ -52,37 +47,33 @@ use crate::{input, renderer};
 pub struct WindowContext {
     pub message_buffer: MessageBuffer,
     pub display: Display,
-    event_queue: Vec<WinitEvent<'static, Event>>,
+    pub dirty: bool,
+    event_queue: Vec<WinitEvent<Event>>,
     terminal: Arc<FairMutex<Term<EventProxy>>>,
     cursor_blink_timed_out: bool,
-    modifiers: ModifiersState,
+    modifiers: Modifiers,
+    inline_search_state: InlineSearchState,
     search_state: SearchState,
-    received_count: usize,
-    suppress_chars: bool,
     notifier: Notifier,
-    font_size: Size,
     mouse: Mouse,
     touch: TouchPurpose,
-    dirty: bool,
     occluded: bool,
     preserve_title: bool,
     #[cfg(not(windows))]
     master_fd: RawFd,
     #[cfg(not(windows))]
     shell_pid: u32,
-    ipc_config: Vec<(String, serde_yaml::Value)>,
+    window_config: ParsedOptions,
     config: Rc<UiConfig>,
 }
 
 impl WindowContext {
-    /// Create initial window context that dous bootstrapping the graphics Api we're going to use.
+    /// Create initial window context that does bootstrapping the graphics API we're going to use.
     pub fn initial(
         event_loop: &EventLoopWindowTarget<Event>,
         proxy: EventLoopProxy<Event>,
         config: Rc<UiConfig>,
         options: WindowOptions,
-        #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-        wayland_event_queue: Option<&EventQueue>,
     ) -> Result<Self, Box<dyn Error>> {
         let raw_display_handle = event_loop.raw_display_handle();
 
@@ -99,8 +90,11 @@ impl WindowContext {
         #[cfg(not(windows))]
         let raw_window_handle = None;
 
-        let gl_display =
-            renderer::platform::create_gl_display(raw_display_handle, raw_window_handle)?;
+        let gl_display = renderer::platform::create_gl_display(
+            raw_display_handle,
+            raw_window_handle,
+            config.debug.prefer_egl,
+        )?;
         let gl_config = renderer::platform::pick_gl_config(&gl_display, raw_window_handle)?;
 
         #[cfg(not(windows))]
@@ -108,17 +102,19 @@ impl WindowContext {
             event_loop,
             &config,
             &identity,
-            #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-            wayland_event_queue,
             #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
             gl_config.x11_visual(),
+            #[cfg(target_os = "macos")]
+            &options.window_tabbing_id,
         )?;
 
         // Create context.
         let gl_context =
             renderer::platform::create_gl_context(&gl_display, &gl_config, raw_window_handle)?;
 
-        Self::new(window, gl_context, config, options, proxy)
+        let display = Display::new(window, gl_context, &config, false)?;
+
+        Self::new(display, config, options, proxy)
     }
 
     /// Create additional context with the graphics platform other windows are using.
@@ -128,8 +124,7 @@ impl WindowContext {
         proxy: EventLoopProxy<Event>,
         config: Rc<UiConfig>,
         options: WindowOptions,
-        #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-        wayland_event_queue: Option<&EventQueue>,
+        config_overrides: ParsedOptions,
     ) -> Result<Self, Box<dyn Error>> {
         // Get any window and take its GL config and display to build a new context.
         let (gl_display, gl_config) = {
@@ -144,10 +139,10 @@ impl WindowContext {
             event_loop,
             &config,
             &identity,
-            #[cfg(all(feature = "wayland", not(any(target_os = "macos", windows))))]
-            wayland_event_queue,
             #[cfg(all(feature = "x11", not(any(target_os = "macos", windows))))]
             gl_config.x11_visual(),
+            #[cfg(target_os = "macos")]
+            &options.window_tabbing_id,
         )?;
 
         // Create context.
@@ -158,26 +153,35 @@ impl WindowContext {
             Some(raw_window_handle),
         )?;
 
-        Self::new(window, gl_context, config, options, proxy)
+        // Check if new window will be opened as a tab.
+        #[cfg(target_os = "macos")]
+        let tabbed = options.window_tabbing_id.is_some();
+        #[cfg(not(target_os = "macos"))]
+        let tabbed = false;
+
+        let display = Display::new(window, gl_context, &config, tabbed)?;
+
+        let mut window_context = Self::new(display, config, options, proxy)?;
+
+        // Set the config overrides at startup.
+        //
+        // These are already applied to `config`, so no update is necessary.
+        window_context.window_config = config_overrides;
+
+        Ok(window_context)
     }
 
     /// Create a new terminal window context.
     fn new(
-        window: Window,
-        context: NotCurrentContext,
+        display: Display,
         config: Rc<UiConfig>,
         options: WindowOptions,
         proxy: EventLoopProxy<Event>,
     ) -> Result<Self, Box<dyn Error>> {
-        let mut pty_config = config.terminal_config.pty_config.clone();
+        let mut pty_config = config.pty_config();
         options.terminal_options.override_pty_config(&mut pty_config);
 
         let preserve_title = options.window_identity.title.is_some();
-
-        // Create a display.
-        //
-        // The display manages a window and can draw the terminal.
-        let display = Display::new(window, context, &config)?;
 
         info!(
             "PTY dimensions: {:?} x {:?}",
@@ -192,7 +196,7 @@ impl WindowContext {
         // This object contains all of the state about what's being displayed. It's
         // wrapped in a clonable mutex since both the I/O loop and display need to
         // access it.
-        let terminal = Term::new(&config.terminal_config, &display.size_info, event_proxy.clone());
+        let terminal = Term::new(config.term_options(), &display.size_info, event_proxy.clone());
         let terminal = Arc::new(FairMutex::new(terminal));
 
         // Create the PTY.
@@ -219,7 +223,7 @@ impl WindowContext {
             pty,
             pty_config.hold,
             config.debug.ref_test,
-        );
+        )?;
 
         // The event loop channel allows write requests from the event processor
         // to be sent to the pty loop and ultimately written to the pty.
@@ -229,16 +233,13 @@ impl WindowContext {
         let _io_thread = event_loop.spawn();
 
         // Start cursor blinking, in case `Focused` isn't sent on startup.
-        if config.terminal_config.cursor.style().blinking {
+        if config.cursor.style().blinking {
             event_proxy.send_event(TerminalEvent::CursorBlinkingChange.into());
         }
-
-        let font_size = config.font.size();
 
         // Create context for the Alacritty window.
         Ok(WindowContext {
             preserve_title,
-            font_size,
             terminal,
             display,
             #[cfg(not(windows))]
@@ -248,17 +249,16 @@ impl WindowContext {
             config,
             notifier: Notifier(loop_tx),
             cursor_blink_timed_out: Default::default(),
-            suppress_chars: Default::default(),
+            inline_search_state: Default::default(),
             message_buffer: Default::default(),
-            received_count: Default::default(),
+            window_config: Default::default(),
             search_state: Default::default(),
             event_queue: Default::default(),
-            ipc_config: Default::default(),
             modifiers: Default::default(),
+            occluded: Default::default(),
             mouse: Default::default(),
             touch: Default::default(),
             dirty: Default::default(),
-            occluded: Default::default(),
         })
     }
 
@@ -267,50 +267,29 @@ impl WindowContext {
         let old_config = mem::replace(&mut self.config, new_config);
 
         // Apply ipc config if there are overrides.
-        if !self.ipc_config.is_empty() {
-            let mut config = (*self.config).clone();
-
-            // Apply each option, removing broken ones.
-            let mut i = 0;
-            while i < self.ipc_config.len() {
-                let (key, value) = &self.ipc_config[i];
-
-                match config.replace(key, value.clone()) {
-                    Err(err) => {
-                        error!(
-                            target: LOG_TARGET_IPC_CONFIG,
-                            "Unable to override option '{}': {}", key, err
-                        );
-                        self.ipc_config.swap_remove(i);
-                    },
-                    Ok(_) => i += 1,
-                }
-            }
-
-            self.config = Rc::new(config);
-        }
+        self.config = self.window_config.override_config_rc(self.config.clone());
 
         self.display.update_config(&self.config);
-        self.terminal.lock().update_config(&self.config.terminal_config);
+        self.terminal.lock().set_options(self.config.term_options());
 
         // Reload cursor if its thickness has changed.
-        if (old_config.terminal_config.cursor.thickness()
-            - self.config.terminal_config.cursor.thickness())
-        .abs()
-            > f32::EPSILON
-        {
+        if (old_config.cursor.thickness() - self.config.cursor.thickness()).abs() > f32::EPSILON {
             self.display.pending_update.set_cursor_dirty();
         }
 
         if old_config.font != self.config.font {
+            let scale_factor = self.display.window.scale_factor as f32;
             // Do not update font size if it has been changed at runtime.
-            if self.font_size == old_config.font.size() {
-                self.font_size = self.config.font.size();
+            if self.display.font_size == old_config.font.size().scale(scale_factor) {
+                self.display.font_size = self.config.font.size().scale(scale_factor);
             }
 
-            let font = self.config.font.clone().with_size(self.font_size);
+            let font = self.config.font.clone().with_size(self.display.font_size);
             self.display.pending_update.set_font(font);
         }
+
+        // Always reload the theme to account for auto-theme switching.
+        self.display.window.set_theme(self.config.window.theme());
 
         // Update display if either padding options or resize increments were changed.
         let window_config = &old_config.window;
@@ -342,10 +321,11 @@ impl WindowContext {
         self.display.window.set_has_shadow(opaque);
 
         #[cfg(target_os = "macos")]
-        self.display.window.set_option_as_alt(self.config.window.option_as_alt);
+        self.display.window.set_option_as_alt(self.config.window.option_as_alt());
 
-        // Change opacity state.
+        // Change opacity and blur state.
         self.display.window.set_transparent(!opaque);
+        self.display.window.set_blur(self.config.window.blur);
 
         // Update hint keys.
         self.display.hint_state.update_alphabet(self.config.hints.alphabet());
@@ -357,74 +337,86 @@ impl WindowContext {
         self.dirty = true;
     }
 
-    /// Update the IPC config overrides.
+    /// Clear the window config overrides.
     #[cfg(unix)]
-    pub fn update_ipc_config(&mut self, config: Rc<UiConfig>, ipc_config: IpcConfig) {
-        // Clear previous IPC errors.
+    pub fn reset_window_config(&mut self, config: Rc<UiConfig>) {
+        // Clear previous window errors.
         self.message_buffer.remove_target(LOG_TARGET_IPC_CONFIG);
 
-        if ipc_config.reset {
-            self.ipc_config.clear();
-        } else {
-            for option in &ipc_config.options {
-                // Separate config key/value.
-                let (key, value) = match option.split_once('=') {
-                    Some(split) => split,
-                    None => {
-                        error!(
-                            target: LOG_TARGET_IPC_CONFIG,
-                            "'{}': IPC config option missing value", option
-                        );
-                        continue;
-                    },
-                };
-
-                // Try and parse value as yaml.
-                match serde_yaml::from_str(value) {
-                    Ok(value) => self.ipc_config.push((key.to_owned(), value)),
-                    Err(err) => error!(
-                        target: LOG_TARGET_IPC_CONFIG,
-                        "'{}': Invalid IPC config value: {:?}", option, err
-                    ),
-                }
-            }
-        }
+        self.window_config.clear();
 
         // Reload current config to pull new IPC config.
         self.update_config(config);
     }
 
+    /// Add new window config overrides.
+    #[cfg(unix)]
+    pub fn add_window_config(&mut self, config: Rc<UiConfig>, options: &ParsedOptions) {
+        // Clear previous window errors.
+        self.message_buffer.remove_target(LOG_TARGET_IPC_CONFIG);
+
+        self.window_config.extend_from_slice(options);
+
+        // Reload current config to pull new IPC config.
+        self.update_config(config);
+    }
+
+    /// Draw the window.
+    pub fn draw(&mut self, scheduler: &mut Scheduler) {
+        self.display.window.requested_redraw = false;
+
+        if self.occluded {
+            return;
+        }
+
+        self.dirty = false;
+
+        // Force the display to process any pending display update.
+        self.display.process_renderer_update();
+
+        // Request immediate re-draw if visual bell animation is not finished yet.
+        if !self.display.visual_bell.completed() {
+            // We can get an OS redraw which bypasses alacritty's frame throttling, thus
+            // marking the window as dirty when we don't have frame yet.
+            if self.display.window.has_frame {
+                self.display.window.request_redraw();
+            } else {
+                self.dirty = true;
+            }
+        }
+
+        // Redraw the window.
+        let terminal = self.terminal.lock();
+        self.display.draw(
+            terminal,
+            scheduler,
+            &self.message_buffer,
+            &self.config,
+            &mut self.search_state,
+        );
+    }
+
     /// Process events for this terminal window.
     pub fn handle_event(
         &mut self,
-        event_loop: &EventLoopWindowTarget<Event>,
+        #[cfg(target_os = "macos")] event_loop: &EventLoopWindowTarget<Event>,
         event_proxy: &EventLoopProxy<Event>,
         clipboard: &mut Clipboard,
         scheduler: &mut Scheduler,
-        event: WinitEvent<'_, Event>,
+        event: WinitEvent<Event>,
     ) {
         match event {
-            // Skip further event handling with no staged updates.
-            WinitEvent::RedrawEventsCleared if self.event_queue.is_empty() && !self.dirty => {
-                return;
+            WinitEvent::AboutToWait
+            | WinitEvent::WindowEvent { event: WindowEvent::RedrawRequested, .. } => {
+                // Skip further event handling with no staged updates.
+                if self.event_queue.is_empty() {
+                    return;
+                }
+
+                // Continue to process all pending events.
             },
-            // Continue to process all pending events.
-            WinitEvent::RedrawEventsCleared => (),
-            // Remap scale_factor change event to remove the lifetime.
-            WinitEvent::WindowEvent {
-                event: WindowEvent::ScaleFactorChanged { scale_factor, new_inner_size },
-                window_id,
-            } => {
-                let size = (new_inner_size.width, new_inner_size.height);
-                let event =
-                    Event::new(EventType::ScaleFactorChanged(scale_factor, size), window_id);
-                self.event_queue.push(event.into());
-                return;
-            },
-            // Transmute to extend lifetime, which exists only for `ScaleFactorChanged` event.
-            // Since we remap that event to remove the lifetime, this is safe.
-            event => unsafe {
-                self.event_queue.push(mem::transmute(event));
+            event => {
+                self.event_queue.push(event);
                 return;
             },
         }
@@ -436,11 +428,9 @@ impl WindowContext {
         let context = ActionContext {
             cursor_blink_timed_out: &mut self.cursor_blink_timed_out,
             message_buffer: &mut self.message_buffer,
-            received_count: &mut self.received_count,
-            suppress_chars: &mut self.suppress_chars,
+            inline_search_state: &mut self.inline_search_state,
             search_state: &mut self.search_state,
             modifiers: &mut self.modifiers,
-            font_size: &mut self.font_size,
             notifier: &mut self.notifier,
             display: &mut self.display,
             mouse: &mut self.mouse,
@@ -455,6 +445,7 @@ impl WindowContext {
             preserve_title: self.preserve_title,
             config: &self.config,
             event_proxy,
+            #[cfg(target_os = "macos")]
             event_loop,
             clipboard,
             scheduler,
@@ -472,7 +463,7 @@ impl WindowContext {
                 &mut self.display,
                 &mut self.notifier,
                 &self.message_buffer,
-                &self.search_state,
+                &mut self.search_state,
                 old_is_searching,
                 &self.config,
             );
@@ -484,35 +475,19 @@ impl WindowContext {
                 &terminal,
                 &self.config,
                 &self.mouse,
-                self.modifiers,
+                self.modifiers.state(),
             );
             self.mouse.hint_highlight_dirty = false;
         }
 
-        // Skip rendering until we get a new frame.
-        if !self.display.window.has_frame.load(Ordering::Relaxed) {
-            return;
-        }
-
-        if self.dirty && !self.occluded {
-            // Force the display to process any pending display update.
-            self.display.process_renderer_update();
-
-            self.dirty = false;
-
-            // Request immediate re-draw if visual bell animation is not finished yet.
-            if !self.display.visual_bell.completed() {
-                self.display.window.request_redraw();
-            }
-
-            // Redraw the window.
-            self.display.draw(
-                terminal,
-                scheduler,
-                &self.message_buffer,
-                &self.config,
-                &self.search_state,
-            );
+        // Don't call `request_redraw` when event is `RedrawRequested` since the `dirty` flag
+        // represents the current frame, but redraw is for the next frame.
+        if self.dirty
+            && self.display.window.has_frame
+            && !self.occluded
+            && !matches!(event, WinitEvent::WindowEvent { event: WindowEvent::RedrawRequested, .. })
+        {
+            self.display.window.request_redraw();
         }
     }
 
@@ -555,7 +530,7 @@ impl WindowContext {
         display: &mut Display,
         notifier: &mut Notifier,
         message_buffer: &MessageBuffer,
-        search_state: &SearchState,
+        search_state: &mut SearchState,
         old_is_searching: bool,
         config: &UiConfig,
     ) {
@@ -568,13 +543,7 @@ impl WindowContext {
             search_state.direction == Direction::Left
         };
 
-        display.handle_update(
-            terminal,
-            notifier,
-            message_buffer,
-            search_state.history_index.is_some(),
-            config,
-        );
+        display.handle_update(terminal, notifier, message_buffer, search_state, config);
 
         let new_is_searching = search_state.history_index.is_some();
         if !old_is_searching && new_is_searching {
