@@ -1,6 +1,7 @@
 use std::cmp::min;
 
-use alacritty_config_derive::ConfigDeserialize;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 use crate::event::EventListener;
 use crate::grid::{Dimensions, GridCell};
@@ -9,7 +10,8 @@ use crate::term::cell::Flags;
 use crate::term::Term;
 
 /// Possible vi mode motion movements.
-#[derive(ConfigDeserialize, Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize), serde(rename_all = "lowercase"))]
 pub enum ViMotion {
     /// Move up.
     Up,
@@ -19,11 +21,12 @@ pub enum ViMotion {
     Left,
     /// Move right.
     Right,
-    /// Move to start of line.
+    /// First column, or beginning of the line when already at the first column.
     First,
-    /// Move to end of line.
+    /// Last column, or beginning of the line when already at the last column.
     Last,
-    /// Move to the first non-empty cell.
+    /// First non-empty cell in this terminal row, or first non-empty cell
+    /// of the line when already at the first cell of the row.
     FirstOccupied,
     /// Move to top of screen.
     High,
@@ -242,7 +245,7 @@ fn first_occupied<T>(term: &Term<T>, mut point: Point) -> Point {
 
 /// Move by semantically separated word, like w/b/e/ge in vi.
 fn semantic<T: EventListener>(
-    term: &mut Term<T>,
+    term: &Term<T>,
     mut point: Point,
     direction: Direction,
     side: Side,
@@ -292,7 +295,7 @@ fn semantic<T: EventListener>(
 
 /// Move by whitespace separated word, like W/B/E/gE in vi.
 fn word<T: EventListener>(
-    term: &mut Term<T>,
+    term: &Term<T>,
     mut point: Point,
     direction: Direction,
     side: Side,
@@ -378,16 +381,15 @@ fn is_boundary<T>(term: &Term<T>, point: Point, direction: Direction) -> bool {
 mod tests {
     use super::*;
 
-    use crate::ansi::Handler;
-    use crate::config::Config;
     use crate::event::VoidListener;
     use crate::index::{Column, Line};
     use crate::term::test::TermSize;
-    use crate::term::Term;
+    use crate::term::{Config, Term};
+    use crate::vte::ansi::Handler;
 
     fn term() -> Term<VoidListener> {
         let size = TermSize::new(20, 20);
-        Term::new(&Config::default(), &size, VoidListener)
+        Term::new(Config::default(), &size, VoidListener)
     }
 
     #[test]
